preprocessor/graph_db/requirements.txt
base58==2.1.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cytoolz==0.12.0
ecdsa==0.18.0
eth-hash==0.5.1
eth-keys==0.4.0
eth-typing==3.2.0
eth-utils==2.1.0
greenlet==2.0.1
idna==3.4
interchange==2021.0.4
monotonic==1.6
more-itertools==9.0.0
packaging==21.3
pansi==2020.7.3
psycopg2-binary==2.9.5
py-bip39-bindings==0.1.10
py-ed25519-zebra-bindings==1.0.1
py-sr25519-bindings==0.2.0
py2neo==2021.2.3
pycparser==2.21
pycryptodome==3.15.0
Pygments==2.13.0
PyNaCl==1.5.0
pyparsing==3.0.9
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
scalecodec==1.0.48
six==1.16.0
SQLAlchemy==1.4.44
substrate-interface==1.3.4
substrateinterface==1.0.0
toolz==0.12.0
urllib3==1.26.12
websocket-client==1.4.2
xxhash==3.1.0
